moved TriggerMatcher to its own header
... to be reusable by other block implementations.
